+++ flutter_local_notifications/lib/src/platform_specifics/android/notification_details.dart
@@ -334,8 +334,9 @@ class AndroidNotificationDetails {
   /// soon as it triggers.
   ///
   /// Note: The system UI may choose to display a heads-up notification,
-  /// instead of launching your full-screen intent, while the user is using the
-  /// device. When the full-screen intent occurs, the plugin will act as though
+  /// instead of launching your full-screen intent. This can occur while the
+  /// user is using the device or due the full-screen intent not being granted.
+  /// When the full-screen intent occurs, the plugin will act as though
   /// the user has tapped on a notification so handle it the same way
   /// (e.g. via `onSelectNotification` callback) to display the appropriate
   /// page for your application.