+++ flutter_local_notifications/test/android_flutter_local_notifications_test.dart
@@ -2029,6 +2029,140 @@ void main() {
       }
     });
 
+    group('periodicallyShowWithDuration', () {
+      final DateTime now = DateTime(2023, 12, 29);
+
+      const Duration thirtySeconds = Duration(seconds: 30);
+      test('$thirtySeconds', () async {
+        await withClock(Clock.fixed(now), () async {
+          const AndroidInitializationSettings androidInitializationSettings =
+              AndroidInitializationSettings('app_icon');
+          const InitializationSettings initializationSettings =
+              InitializationSettings(android: androidInitializationSettings);
+          await flutterLocalNotificationsPlugin
+              .initialize(initializationSettings);
+
+          const AndroidNotificationDetails androidNotificationDetails =
+              AndroidNotificationDetails('channelId', 'channelName',
+                  channelDescription: 'channelDescription');
+
+          expect(
+              () async => await flutterLocalNotificationsPlugin
+                      .periodicallyShowWithDuration(
+                    1,
+                    'notification title',
+                    'notification body',
+                    thirtySeconds,
+                    const NotificationDetails(
+                        android: androidNotificationDetails),
+                  ),
+              throwsA(isA<ArgumentError>()));
+        });
+      });
+
+      final List<Duration> repeatDurationIntervals = <Duration>[
+        const Duration(minutes: 1),
+        const Duration(minutes: 15),
+        const Duration(hours: 5),
+        const Duration(days: 30)
+      ];
+
+      for (final Duration repeatDurationInterval in repeatDurationIntervals) {
+        test('$repeatDurationInterval', () async {
+          await withClock(Clock.fixed(now), () async {
+            const AndroidInitializationSettings androidInitializationSettings =
+                AndroidInitializationSettings('app_icon');
+            const InitializationSettings initializationSettings =
+                InitializationSettings(android: androidInitializationSettings);
+            await flutterLocalNotificationsPlugin
+                .initialize(initializationSettings);
+
+            const AndroidNotificationDetails androidNotificationDetails =
+                AndroidNotificationDetails('channelId', 'channelName',
+                    channelDescription: 'channelDescription');
+            await flutterLocalNotificationsPlugin.periodicallyShowWithDuration(
+              1,
+              'notification title',
+              'notification body',
+              repeatDurationInterval,
+              const NotificationDetails(android: androidNotificationDetails),
+            );
+
+            expect(
+                log.last,
+                isMethodCall('periodicallyShowWithDuration',
+                    arguments: <String, Object>{
+                      'id': 1,
+                      'title': 'notification title',
+                      'body': 'notification body',
+                      'payload': '',
+                      'calledAt': now.millisecondsSinceEpoch,
+                      'repeatIntervalMilliseconds':
+                          repeatDurationInterval.inMilliseconds,
+                      'platformSpecifics': <String, Object?>{
+                        'scheduleMode': 'exact',
+                        'icon': null,
+                        'channelId': 'channelId',
+                        'channelName': 'channelName',
+                        'channelDescription': 'channelDescription',
+                        'channelShowBadge': true,
+                        'channelAction': AndroidNotificationChannelAction
+                            .createIfNotExists.index,
+                        'importance': Importance.defaultImportance.value,
+                        'priority': Priority.defaultPriority.value,
+                        'playSound': true,
+                        'enableVibration': true,
+                        'vibrationPattern': null,
+                        'groupKey': null,
+                        'setAsGroupSummary': false,
+                        'groupAlertBehavior': GroupAlertBehavior.all.index,
+                        'autoCancel': true,
+                        'ongoing': false,
+                        'silent': false,
+                        'colorAlpha': null,
+                        'colorRed': null,
+                        'colorGreen': null,
+                        'colorBlue': null,
+                        'onlyAlertOnce': false,
+                        'showWhen': true,
+                        'when': null,
+                        'usesChronometer': false,
+                        'chronometerCountDown': false,
+                        'showProgress': false,
+                        'maxProgress': 0,
+                        'progress': 0,
+                        'indeterminate': false,
+                        'enableLights': false,
+                        'ledColorAlpha': null,
+                        'ledColorRed': null,
+                        'ledColorGreen': null,
+                        'ledColorBlue': null,
+                        'ledOnMs': null,
+                        'ledOffMs': null,
+                        'ticker': null,
+                        'visibility': null,
+                        'timeoutAfter': null,
+                        'category': null,
+                        'additionalFlags': null,
+                        'fullScreenIntent': false,
+                        'shortcutId': null,
+                        'subText': null,
+                        'style': AndroidNotificationStyle.defaultStyle.index,
+                        'styleInformation': <String, Object>{
+                          'htmlFormatContent': false,
+                          'htmlFormatTitle': false,
+                        },
+                        'tag': null,
+                        'colorized': false,
+                        'number': null,
+                        'audioAttributesUsage': 5,
+                      },
+                    }));
+          });
+        });
+      }
+    });
+
     group('zonedSchedule', () {
       test('no repeat frequency', () async {
         const AndroidInitializationSettings androidInitializationSettings =