+++ flutter_local_notifications/test/macos_flutter_local_notifications_test.dart
@@ -255,6 +255,132 @@ void main() {
       }
     });
 
+    group('periodicallyShowWithDuration', () {
+      final DateTime now = DateTime(2023, 12, 29);
+
+      const Duration thirtySeconds = Duration(seconds: 30);
+      test('$thirtySeconds', () async {
+        await withClock(Clock.fixed(now), () async {
+          const DarwinInitializationSettings macOSInitializationSettings =
+              DarwinInitializationSettings();
+          const InitializationSettings initializationSettings =
+              InitializationSettings(macOS: macOSInitializationSettings);
+          await flutterLocalNotificationsPlugin
+              .initialize(initializationSettings);
+
+          const NotificationDetails notificationDetails = NotificationDetails(
+            macOS: DarwinNotificationDetails(
+              presentAlert: true,
+              presentBadge: true,
+              presentSound: true,
+              presentBanner: true,
+              presentList: true,
+              sound: 'sound.mp3',
+              badgeNumber: 1,
+              attachments: <DarwinNotificationAttachment>[
+                DarwinNotificationAttachment(
+                  'video.mp4',
+                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373',
+                )
+              ],
+            ),
+          );
+
+          expect(
+              () async => await flutterLocalNotificationsPlugin
+                      .periodicallyShowWithDuration(
+                    1,
+                    'notification title',
+                    'notification body',
+                    thirtySeconds,
+                    notificationDetails,
+                  ),
+              throwsA(isA<ArgumentError>()));
+        });
+      });
+
+      final List<Duration> repeatDurationIntervals = <Duration>[
+        const Duration(minutes: 1),
+        const Duration(minutes: 15),
+        const Duration(hours: 5),
+        const Duration(days: 30)
+      ];
+
+      for (final Duration repeatDurationInterval in repeatDurationIntervals) {
+        test('$repeatDurationInterval', () async {
+          await withClock(Clock.fixed(now), () async {
+            const DarwinInitializationSettings macOSInitializationSettings =
+                DarwinInitializationSettings();
+            const InitializationSettings initializationSettings =
+                InitializationSettings(macOS: macOSInitializationSettings);
+            await flutterLocalNotificationsPlugin
+                .initialize(initializationSettings);
+
+            const NotificationDetails notificationDetails = NotificationDetails(
+              macOS: DarwinNotificationDetails(
+                presentAlert: true,
+                presentBadge: true,
+                presentSound: true,
+                presentBanner: true,
+                presentList: true,
+                sound: 'sound.mp3',
+                badgeNumber: 1,
+                attachments: <DarwinNotificationAttachment>[
+                  DarwinNotificationAttachment(
+                    'video.mp4',
+                    identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373',
+                  )
+                ],
+              ),
+            );
+
+            await flutterLocalNotificationsPlugin.periodicallyShowWithDuration(
+              1,
+              'notification title',
+              'notification body',
+              repeatDurationInterval,
+              notificationDetails,
+            );
+
+            expect(
+                log.last,
+                isMethodCall('periodicallyShowWithDuration',
+                    arguments: <String, Object>{
+                      'id': 1,
+                      'title': 'notification title',
+                      'body': 'notification body',
+                      'payload': '',
+                      'calledAt': now.millisecondsSinceEpoch,
+                      'repeatIntervalMilliseconds':
+                          repeatDurationInterval.inMilliseconds,
+                      'platformSpecifics': <String, Object?>{
+                        'presentAlert': true,
+                        'presentBadge': true,
+                        'presentSound': true,
+                        'presentBanner': true,
+                        'presentList': true,
+                        'subtitle': null,
+                        'sound': 'sound.mp3',
+                        'badgeNumber': 1,
+                        'threadIdentifier': null,
+                        'attachments': <Map<String, Object?>>[
+                          <String, Object?>{
+                            'filePath': 'video.mp4',
+                            'identifier':
+                                '2b3f705f-a680-4c9f-8075-a46a70e28373',
+                            'hideThumbnail': null,
+                            'thumbnailClippingRect': null,
+                          }
+                        ],
+                        'categoryIdentifier': null,
+                        'interruptionLevel': null,
+                      },
+                    }));
+          });
+        });
+      }
+    });
+
     group('zonedSchedule', () {
       test('no repeat frequency', () async {
         const DarwinInitializationSettings macOSInitializationSettings =